Microservice/requirements.txt
Flask
Flask-SQLAlchemy
pymysql
cryptography
Flask-Cors
psycopg2-binary
pytest
python-dotenv